backend/requirements.txt
psycopg2-binary==2.8.5
Django~=3.1.0
djangorestframework~=3.11.0
drf-flex-fields~=0.8.6
drf-writable-nested~=0.6.0
django-pdb~=0.6.2
django-extensions~=2.2.9
django-storages~=1.10.1
django-rq~=2.3.2
rqmonitor~=1.0.1
boto3~=1.15.16
Jinja2~=2.11.1
django-watchman~=1.0.0
netaddr~=0.7.19
gunicorn[gevent]~=20.0.4
pyyaml~=5.3.1
uritemplate~=3.0.1
requests~=2.22.0
validators~=0.15.0
django-request-logging~=0.7.2
django-filter~=2.3.0
Werkzeug~=1.0.1
Pillow~=7.2.0
